error resolved maybe

+++ python_async_comprehension/0-async_generator.py
@@ -10,7 +10,7 @@ import random
 from typing import AsyncGenerator
 
 
-async def async_generator() -> AsyncGenerator[int, None]:
+async def async_generator() -> AsyncGenerator[float, None]:
     '''
     Async Generator
     '''
